+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 * [DGPB-2387] Business Rules and attributes "for all" management
 * [DGPB-2445] HDFS file metadata is indexed as an empty table
 * [DGPB-2447] Metrics endpoint for Prometheus
-* [DGPB-2446] Showing log info timing in indexation 
+* [DGPB-2446] Showing log info timing in indexation
+* [DGPB-2468] Searcher with Generic QR. Only index "father" QR
 
 ## 1.1.0-f2011be (Built: July 12, 2019 | Released: July 26, 2019)